+++ relay-filter/src/common.rs
@@ -98,6 +98,12 @@ impl<'de> Deserialize<'de> for GlobPatterns {
     }
 }
 
+impl PartialEq for GlobPatterns {
+    fn eq(&self, other: &Self) -> bool {
+        self.patterns == other.patterns
+    }
+}
+
 /// Identifies which filter dropped an event for which reason.
 ///
 /// Ported from Sentry's same-named "enum". The enum variants are fed into outcomes in kebap-case
@@ -127,6 +133,9 @@ pub enum FilterStatKey {
 
     /// Filtered due to invalid CSP policy.
     InvalidCsp,
+
+    /// Filtered due to the fact that it was a call to a filtered transaction
+    FilteredTransactions,
 }
 
 // An event grouped to a removed group.
@@ -153,6 +162,7 @@ impl FilterStatKey {
             FilterStatKey::Localhost => "localhost",
             FilterStatKey::WebCrawlers => "web-crawlers",
             FilterStatKey::InvalidCsp => "invalid-csp",
+            FilterStatKey::FilteredTransactions => "filtered-transaction",
         }
     }
 }
@@ -176,6 +186,7 @@ impl<'a> TryFrom<&'a str> for FilterStatKey {
             "localhost" => FilterStatKey::Localhost,
             "web-crawlers" => FilterStatKey::WebCrawlers,
             "invalid-csp" => FilterStatKey::InvalidCsp,
+            "filtered-transaction" => FilterStatKey::FilteredTransactions,
             other => {
                 return Err(other);
             }